refactor: 패키지 이동, 인수테스트 클래스명 수정
