+++ backend/src/main/java/com/funeat/review/domain/Review.java
@@ -3,7 +3,6 @@ package com.funeat.review.domain;
 import com.funeat.member.domain.Member;
 import com.funeat.member.domain.favorite.ReviewFavorite;
 import com.funeat.product.domain.Product;
-import java.util.ArrayList;
 import java.util.List;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -37,7 +36,7 @@ public class Review {
     private Member member;
 
     @OneToMany(mappedBy = "review")
-    private List<ReviewTag> reviewTags = new ArrayList<>();
+    private List<ReviewTag> reviewTags;
 
     @OneToMany(mappedBy = "review")
     private List<ReviewFavorite> reviewFavorites;
